Split commit rs into multiple files
This single file was getting unweildy, so I ended up splitting it up into multiple files for ease of use of use.

It also updates commit_move to use the correct signature
