+++ crates/but-api/src/commit/mod.rs
@@ -1,0 +1,26 @@
+/// Functions for amending existing commits.
+pub mod amend;
+
+/// Functions for creating new commits.
+pub mod create;
+
+/// JSON transport types for commit APIs.
+pub mod json;
+
+/// Functions for inserting blank commits.
+pub mod insert_blank;
+
+/// Functions for moving changes between commits.
+pub mod move_changes;
+
+/// Functions for moving commits within the graph.
+pub mod move_commit;
+
+/// Functions for rewording commits.
+pub mod reword;
+
+/// Shared result and selector types for commit APIs.
+pub mod types;
+
+/// Functions for uncommitting changes from commits.
+pub mod uncommit;
