+++ crates/but/src/command/legacy/commit.rs
@@ -3,7 +3,7 @@ use std::{collections::BTreeMap, fmt::Write as _};
 use anyhow::{Context, Result, bail};
 use bstr::{BString, ByteSlice};
 use but_api::{
-    commit::{RelativeTo, commit_create, commit_insert_blank},
+    commit::{create::commit_create, insert_blank::commit_insert_blank, types::RelativeTo},
     diff,
     legacy::{repo, workspace},
 };
